move `intersection` to `set-utils`
